Type들을 관리하기 위한 Types.swift 작성
